requirements.txt
py-cord==2.0.0b4
python-dotenv==0.20.0
